requirements.txt
openai==1.54.5
elevenlabs==1.11.0
python-dotenv==1.0.1
